requirements.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
firebase-admin==6.1.0
google-cloud-storage==2.7.0